[security] implemented token checks for all sensitve endpoints in users and appointments

+++ backend/app/routes/appointment_router.py
@@ -5,7 +5,6 @@ from datetime import datetime
 from typing import Annotated
 
 from fastapi import APIRouter, Depends, HTTPException, status, Response
-from fastapi.security import OAuth2PasswordBearer
 from sqlmodel import Session, select, extract
 
 from app.db import get_session
@@ -20,36 +19,27 @@ from app.services.distance_service import get_or_create_distance
 from app.services.pdf_service import generate_pdf_report
 from app.core.security import get_current_user
 
-appointment_router = APIRouter(tags=["appointments"])
+
+appointment_router = APIRouter(prefix='/appointments', tags=["appointments"], )
 
 
-@appointment_router.post(
-    "/users/{user_id}/appointments",
-    response_model=AppointmentPublic,
-    status_code=status.HTTP_201_CREATED,
-)
+@appointment_router.post("/", response_model=AppointmentPublic)
 def create_appointment(
-    user_id: int,
     appointment_data: AppointmentCreate,
+    current_user: Annotated[User, Depends(get_current_user)],
     session: Session = Depends(get_session),
 ):
-    user = session.get(User, user_id)
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
-
-
+    """
+    Creates an appointment by taking in an AppointmentCreate object 
+    """
     distance = get_or_create_distance(
         session=session,
-        origin_address=user.address,
+        origin_address=current_user.address,
         destination_address=appointment_data.destination_address,
     )
 
     extra_data = {
-        'user_id': user.id,
+        'user_id': current_user.id,
         'distance_id': distance.id,
         'roundtrip_distance': distance.roundtrip_distance
     }
@@ -63,11 +53,25 @@ def create_appointment(
     return new_appointment
 
 
-@appointment_router.get("/appointments/{appointment_id}", response_model=AppointmentPublic)
-def get_appointment(
-    appointment_id: int,
+@appointment_router.get('/', response_model=list[AppointmentPublic])
+def get_appointment_list(
+    current_user: Annotated[User, Depends(get_current_user)],
     session: Session = Depends(get_session),
 ):
+    appointments = session.exec(
+        select(Appointment)
+        .where(Appointment.user_id == current_user.id)
+    ).all()
+
+    return appointments
+
+
+@appointment_router.get("/{appointment_id}", response_model=AppointmentPublic)
+def get_appointment(
+    appointment_id: int,
+    current_user: Annotated[User, Depends(get_current_user)],
+    session: Session = Depends(get_session),
+):    
     appointment = session.get(Appointment, appointment_id)
 
     if not appointment:
@@ -75,45 +79,44 @@ def get_appointment(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Appointment not found",
         )
+    
+    if appointment.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail='Unauthorized access'
+            )
 
     return appointment
 
 
-@appointment_router.get("/users/{user_id}/appointments/reports")
+@appointment_router.get("/reports")
 def get_user_appointments_report(
     current_user: Annotated[User, Depends(get_current_user)],
-    user_id: int,
     year: int | None = None,
     session: Session = Depends(get_session),
 ) -> Response:
     report_year = year or datetime.now().year
-    user = session.get(User, user_id)
-    
-    if current_user.id != user_id:
-        raise HTTPException(
-            status_code=status.HTTP_403_FORBIDDEN,
-            detail='You do not have permission to acces this user\'s reports'
-        )
 
     appointments = session.exec(
         select(Appointment)
-        .where(Appointment.user_id == user_id)
-        .where(extract('year', Appointment.appointment_date) == year)
+        .where(Appointment.user_id == current_user.id)
+        .where(extract('year', Appointment.appointment_date) == report_year)
         .order_by(Appointment.appointment_date, Appointment.id)
     ).all()
 
-    if len(appointments) == 0:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'No appointments found for {year}')
+    if not appointments:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'No appointments found for {report_year}')
     
-    return generate_pdf_report(appointments, user, report_year)
+    return generate_pdf_report(appointments, current_user, report_year)
 
 
-@appointment_router.patch("/appointments/{appointment_id}", response_model=AppointmentPublic)
+@appointment_router.patch("/{appointment_id}", response_model=AppointmentPublic)
 def update_appointment(
     appointment_id: int,
     appointment_data: AppointmentUpdate,
+    current_user: Annotated[User, Depends(get_current_user)],
     session: Session = Depends(get_session),
-):
+):   
     appointment = session.get(Appointment, appointment_id)
 
     if not appointment:
@@ -121,26 +124,23 @@ def update_appointment(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Appointment not found",
         )
+    
+    if appointment.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail='Unauthorized access'
+            )
 
     appointment_data_dump = appointment_data.model_dump(exclude_unset=True)
 
-    if "destination_address" in appointment_data_dump:
-        user = session.get(User, appointment.user_id)
+    distance = get_or_create_distance(
+        session=session,
+        origin_address=current_user.address,
+        destination_address=appointment_data_dump["destination_address"],
+    )
 
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Associated user not found",
-            )
-
-        distance = get_or_create_distance(
-            session=session,
-            origin_address=user.address,
-            destination_address=appointment_data_dump["destination_address"],
-        )
-
-        appointment.distance_id = distance.id
-        appointment.roundtrip_distance = distance.roundtrip_distance
+    appointment.distance_id = distance.id
+    appointment.roundtrip_distance = distance.roundtrip_distance
 
     appointment.sqlmodel_update(appointment_data_dump)
 
@@ -151,23 +151,21 @@ def update_appointment(
     return appointment
 
 
-@appointment_router.delete("/appointments/{appointment_id}")
+@appointment_router.delete("/{appointment_id}")
 def delete_appointment(
     appointment_id: int,
+    current_user: Annotated[User, Depends(get_current_user)],
     session: Session = Depends(get_session),
 ):
     appointment = session.get(Appointment, appointment_id)
 
     if not appointment:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Appointment not found",
-        )
+        raise HTTPException(status_code=404, detail="Appointment not found")
+    
+    if appointment.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail='Access forbidden')
 
     session.delete(appointment)
     session.commit()
 
-    return {
-        "message": "Appointment deleted",
-        "appointment_id": appointment_id,
-    }
+    return {"message": "Appointment deleted", "appointment_id": appointment_id}
